backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
    "bcrypt==3.2.0",
    "opencv-python-headless>=4.13.0.92",
    "imagehash>=4.3.2",
    "pillow>=12.2.0",
    "pdqhash>=0.2.8",
    "yt-dlp>=2026.3.17",
    "tenacity>=9.1.4",
    "google-api-python-client>=2.194.0",
    "tavily-python>=0.7.23",
    "requests>=2.33.1",
    "langchain-ollama>=1.1.0",
]
